Implement Suggestion For Clothes

+++ app/src/main/java/com/example/clothessuggester/model/Current.kt
@@ -1,7 +1,7 @@
 package com.example.clothessuggester.model
 
 data class Current(
-    val temperature: Double,
+    val temperature: Int,
     val weather_descriptions: List<String>,
     val is_day: String
 )
